pyproject.toml
[tool.poetry]
name = "serialmidi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
python-rtmidi = "^1.4.3"
PyInquirer = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
